requirements.txt
# requirements.txt 

PyMuPDF==1.23.5
PyPDF2==3.0.1
python-docx==1.1.2
lxml==5.2.2